pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
dependencies = [
    "mysql-connector-python>=8.0.0",
    "configparser>=5.0.0",
    "sshtunnel>=0.4.0",
    "paramiko<4",
]
name = "guacalib"
description = "Guacamole management library and CLI"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name="Roman V. Isaev", email="rm@isaeff.net"}
]
license = "GPL-3.0-or-later"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.scripts]
guacaman = "guacalib.cli.main:main"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest",
]

[tool.setuptools.dynamic]
version = {attr = "guacalib.version.VERSION"}
